pyproject.toml
[project]
name = "checking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "google-generativeai>=0.8.5",
    "google-search-results>=2.4.2",
    "openai>=1.88.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
